requirement.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==1.11.14
django-allauth==0.36.0
django-redis==4.10.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
idna==2.7
ipython==5.8.0
ipython-genutils==0.2.0
mysqlclient==1.3.13
oauthlib==2.1.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
python-openid==2.2.5
pytz==2018.5
redis==3.0.1
requests==2.19.1
requests-oauthlib==1.0.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
